@@ -10,17 +10,17 @@
   },
   "dependencies": {
     "@emotion/core": "^11.0.0",
-    "@emotion/react": "^11.11.1",
-    "@emotion/styled": "^11.11.0",
-    "@mui/icons-material": "^5.14.15",
-    "@mui/material": "^5.14.15",
-    "@mui/styled-engine-sc": "^6.0.0-alpha.3",
-    "axios": "^1.5.1",
-    "moment": "^2.29.4",
-    "next": "13.5.6",
+    "@emotion/react": "^11.11.4",
+    "@emotion/styled": "^11.11.5",
+    "@mui/icons-material": "^5.15.19",
+    "@mui/material": "^5.15.19",
+    "@mui/styled-engine-sc": "^6.0.0-alpha.18",
+    "axios": "^1.7.2",
+    "moment": "^2.30.1",
+    "next": "14.2.3",
     "qrcode": "^1.5.3",
     "react": "^18",
     "react-dom": "^18",
-    "styled-components": "^6.1.0"
+    "styled-components": "^6.1.11"
   }
 }
