@@ -1,10 +1,11 @@
+"use client";
 import {
   Alert,
   Box,
   Button,
   IconButton,
   InputAdornment,
-  Typography,
+  Typography
 } from "@mui/material";
 import { styled } from "@mui/material/styles";
 import BoltIcon from "@mui/icons-material/Bolt";
@@ -12,20 +13,20 @@ import QrCode2Icon from "@mui/icons-material/QrCode2";
 import LinkIcon from "@mui/icons-material/Link";
 import TextField from "@mui/material/TextField";
 import ArrowForwardIcon from "@mui/icons-material/ArrowForward";
-import Arrow from "../../../public/arrow.svg";
+import Arrow from "@/public/arrow.svg";
 import Image from "next/image";
 import FiberManualRecordSharpIcon from "@mui/icons-material/FiberManualRecordSharp";
 import TextureSharpIcon from "@mui/icons-material/TextureSharp";
 import Tooltip from "@mui/material/Tooltip";
 import ContentCopyIcon from "@mui/icons-material/ContentCopy";
-import { useEffect, useState } from "react";
+import { useEffect, useRef, useState } from "react";
 import { getShortlyLinkApi } from "@/utils/Api";
 import QRCode from "qrcode";
 import useMediaQuery from "@mui/material/useMediaQuery";
 
 const OutterDiv = styled(Box)(({ theme }) => ({
   display: "flex",
-  justifyContent: "center",
+  justifyContent: "center"
 }));
 
 const InnerDiv = styled(Box)(({ theme }) => ({
@@ -34,13 +35,13 @@ const InnerDiv = styled(Box)(({ theme }) => ({
   width: "85%",
   justifyContent: "space-between",
   [theme.breakpoints.up("lg")]: {
-    width: "1200px",
+    width: "1200px"
   },
   [theme.breakpoints.down("md")]: {
     padding: "10px 20px",
     paddingTop: "80px",
-    width: "100%",
-  },
+    width: "100%"
+  }
 }));
 const MainButton = styled(Button)(({ theme }) => ({
   fontSize: "18px",
@@ -52,16 +53,16 @@ const MainButton = styled(Button)(({ theme }) => ({
   width: "48%",
   letterSpacing: "1px",
   [theme.breakpoints.down("sm")]: {
-    fontSize: "14px",
-  },
+    fontSize: "14px"
+  }
 }));
 const ParentBox = styled(Box)(({ theme }) => ({
   height: "652px",
   position: "relative",
   width: "100%",
   [theme.breakpoints.down("md")]: {
-    height: "720px",
-  },
+    height: "720px"
+  }
 }));
 const MainBox = styled(Box)(({ theme }) => ({
   top: "25%",
@@ -71,8 +72,8 @@ const MainBox = styled(Box)(({ theme }) => ({
   position: "absolute",
   transform: "translate(-50%, -50%)",
   [theme.breakpoints.down("md")]: {
-    top: "20%",
-  },
+    top: "20%"
+  }
 }));
 const MainText = styled(Typography)(({ theme }) => ({
   display: "inline-block",
@@ -84,12 +85,12 @@ const MainText = styled(Typography)(({ theme }) => ({
   textAlign: "center",
   [theme.breakpoints.down("lg")]: {
     fontSize: "64px",
-    lineHeight: "64px",
+    lineHeight: "64px"
   },
   [theme.breakpoints.down("md")]: {
     fontSize: "48px",
-    lineHeight: "48px",
-  },
+    lineHeight: "48px"
+  }
 }));
 
 const SecondText = styled(Typography)(({ theme }) => ({
@@ -102,15 +103,15 @@ const SecondText = styled(Typography)(({ theme }) => ({
   lineHeight: "64px",
   color: "black",
   [theme.breakpoints.down("md")]: {
-    fontSize: "16px",
-  },
+    fontSize: "16px"
+  }
 }));
 
 const ButtonDiv = styled(Box)(({ theme }) => ({
   width: "100%",
   display: "flex",
   flexDirection: "row",
-  justifyContent: "space-between",
+  justifyContent: "space-between"
 }));
 
 function HeroSection() {
@@ -125,6 +126,7 @@ function HeroSection() {
   const thousandWidth = useMediaQuery("(min-width:1000px)");
   const matches = useMediaQuery("(min-width:800px)");
   const sixWidth = useMediaQuery("(min-width:600px)");
+  const inputRef = useRef("");
 
   const InputOutputDiv = styled(Box)(({ theme }) => ({
     left: "50%",
@@ -137,12 +139,12 @@ function HeroSection() {
     gap: "10px",
     position: "absolute",
     [theme.breakpoints.down("md")]: {
-      top: `${linkOrCode === "link" && response ? 48 : 44}%`,
+      top: `${linkOrCode === "link" && response ? 48 : 44}%`
     },
     [theme.breakpoints.down("sm")]: {
       top: `${linkOrCode === "link" && response ? 50 : 46}%`,
-      width: "95%",
-    },
+      width: "95%"
+    }
   }));
 
   const changePreference = (str) => {
@@ -224,7 +226,7 @@ function HeroSection() {
               top: `${matches ? 50 : 25}%`,
               left: "90%",
               color: "#FE7D62",
-              transform: "translate(-50%, -50%) ",
+              transform: "translate(-50%, -50%) "
             }}
           />
           <BoltIcon
@@ -235,7 +237,7 @@ function HeroSection() {
               top: `${matches ? 10 : 5}%`,
               left: `${matches ? 80 : 90}%`,
               color: "#4265F0",
-              transform: "translate(-50%, -50%) rotate(30deg)",
+              transform: "translate(-50%, -50%) rotate(30deg)"
             }}
           />
           <BoltIcon
@@ -246,7 +248,7 @@ function HeroSection() {
               top: `${matches ? 10 : 5}%`,
               left: `${matches ? 20 : 10}%`,
               color: "#FE7D62",
-              transform: "translate(-50%, -50%) rotate(300deg)",
+              transform: "translate(-50%, -50%) rotate(300deg)"
             }}
           />
           {sixWidth && (
@@ -258,7 +260,7 @@ function HeroSection() {
               style={{
                 position: "absolute",
                 top: `${matches ? 49 : 32}%`,
-                left: `${thousandWidth ? 26 : 6}%`,
+                left: `${thousandWidth ? 26 : 6}%`
               }}
             />
           )}
@@ -271,7 +273,7 @@ function HeroSection() {
                 top: "70%",
                 left: "20%",
                 color: "#A1A1A1",
-                transform: "translate(-50%, -50%) rotate(300deg)",
+                transform: "translate(-50%, -50%) rotate(300deg)"
               }}
             />
           )}
@@ -285,7 +287,7 @@ function HeroSection() {
                 top: "60%",
                 left: "10%",
                 color: "#4265F0",
-                transform: "translate(-50%, -50%) rotate(300deg)",
+                transform: "translate(-50%, -50%) rotate(300deg)"
               }}
             />
           )}
@@ -301,7 +303,7 @@ function HeroSection() {
                 flexDirection: "column",
                 justifyContent: "center",
                 alignItems: "center",
-                transform: "translate(-50%, -50%) rotate(330deg)",
+                transform: "translate(-50%, -50%) rotate(330deg)"
               }}
             >
               <Typography
@@ -309,7 +311,7 @@ function HeroSection() {
                   fontSize: "24px",
                   color: "#A1A1A1",
                   letterSpacing: "1px",
-                  fontFamily: "'Poppins', sans-serif",
+                  fontFamily: "'Poppins', sans-serif"
                 }}
               >
                 it's
@@ -319,7 +321,7 @@ function HeroSection() {
                   fontSize: "24px",
                   color: "#A1A1A1",
                   letterSpacing: "1px",
-                  fontFamily: "'Poppins', sans-serif",
+                  fontFamily: "'Poppins', sans-serif"
                 }}
               >
                 free
@@ -330,7 +332,7 @@ function HeroSection() {
           {/* input and output */}
           <InputOutputDiv
             sx={{
-              top: `${linkOrCode === "link" && response ? 58 : 52}%`,
+              top: `${linkOrCode === "link" && response ? 58 : 52}%`
             }}
           >
             <ButtonDiv>
@@ -341,7 +343,7 @@ function HeroSection() {
                   <LinkIcon
                     sx={{
                       transform: "rotate(45deg)",
-                      color: "#e67057",
+                      color: "#e67057"
                     }}
                   />
                 }
@@ -353,15 +355,15 @@ function HeroSection() {
                         backgroundColor: "#4265F0",
                         "&:hover": {
                           color: "#FE7D62",
-                          backgroundColor: "rgba(66, 101, 240, 0.4)",
-                        },
+                          backgroundColor: "rgba(66, 101, 240, 0.4)"
+                        }
                       }
                     : {
                         color: "#4265F0",
                         borderColor: "#4265F0",
                         "&:hover": {
-                          borderColor: "white",
-                        },
+                          borderColor: "white"
+                        }
                       }
                 }
               >
@@ -373,7 +375,7 @@ function HeroSection() {
                 startIcon={
                   <QrCode2Icon
                     sx={{
-                      color: "#FE7D62",
+                      color: "#FE7D62"
                     }}
                   />
                 }
@@ -385,15 +387,15 @@ function HeroSection() {
                         backgroundColor: "#4265F0",
                         "&:hover": {
                           color: "#FE7D62",
-                          backgroundColor: "rgba(66, 101, 240, 0.4)",
-                        },
+                          backgroundColor: "rgba(66, 101, 240, 0.4)"
+                        }
                       }
                     : {
                         color: "#4265F0",
                         borderColor: "#4265F0",
                         "&:hover": {
-                          borderColor: "white",
-                        },
+                          borderColor: "white"
+                        }
                       }
                 }
               >
@@ -409,23 +411,24 @@ function HeroSection() {
             >
               <TextField
                 id="outlined-size-small"
-                onFocus={() => handleFocus()}
-                onBlur={() => handleFocusOut()}
-                value={url}
-                onChange={(e) => {
-                  setShowAlert(false);
-                  setUrl(e.target.value);
-                }}
+                // onFocus={() => handleFocus()}
+                // onBlur={() => handleFocusOut()}
+                inputRef={inputRef}
+                // value={url}
+                // onChange={(e) => {
+                //   // setShowAlert(false);
+                //   setUrl(e.target.value);
+                // }}
                 hiddenLabel
                 sx={{
                   height: "60px",
-                  width: "100%",
+                  width: "100%"
                 }}
                 InputLabelProps={{
                   sx: {
                     color: "#003566",
-                    textTransform: "capitalize",
-                  },
+                    textTransform: "capitalize"
+                  }
                 }}
                 InputProps={{
                   sx: {
@@ -433,9 +436,9 @@ function HeroSection() {
                       // borderRadius: 0,
                     },
                     "&:focus-within fieldset, &:focus-visible fieldset": {
-                      border: "1px solid #FE7D62!important",
+                      border: "1px solid #FE7D62!important"
                       // borderRadius: 0,
-                    },
+                    }
                   },
                   endAdornment: (
                     <InputAdornment
@@ -449,8 +452,8 @@ function HeroSection() {
                         cursor: "pointer",
                         transition: "all 200ms ease-in-out",
                         "&:hover": {
-                          transform: "scale(1.08)",
-                        },
+                          transform: "scale(1.08)"
+                        }
                       }}
                     >
                       <IconButton
@@ -462,20 +465,20 @@ function HeroSection() {
                         sx={{
                           p: "10px",
                           color: "#FE7D62",
-                          fontSize: "14px",
+                          fontSize: "14px"
                         }}
                         aria-label="search"
                       >
                         <ArrowForwardIcon />
                       </IconButton>
                     </InputAdornment>
-                  ),
+                  )
                 }}
                 inputProps={{
                   sx: {
                     color: "#FE7D62",
-                    fontSize: "18px",
-                  },
+                    fontSize: "18px"
+                  }
                 }}
               />
             </form>
@@ -488,14 +491,14 @@ function HeroSection() {
                 value={response}
                 sx={{
                   width: "100%",
-                  height: "60px",
+                  height: "60px"
                   // borderRadius: "0"
                 }}
                 InputLabelProps={{
                   sx: {
                     color: "#003566",
-                    textTransform: "capitalize",
-                  },
+                    textTransform: "capitalize"
+                  }
                 }}
                 InputProps={{
                   sx: {
@@ -503,9 +506,9 @@ function HeroSection() {
                       // borderRadius: 0,
                     },
                     "&:focus-within fieldset, &:focus-visible fieldset": {
-                      border: "1px solid #FE7D62!important",
+                      border: "1px solid #FE7D62!important"
                       // borderRadius: 0,
-                    },
+                    }
                   },
                   endAdornment: (
                     <InputAdornment
@@ -519,8 +522,8 @@ function HeroSection() {
                         cursor: "pointer",
                         transition: "all 200ms ease-in-out",
                         "&:hover": {
-                          transform: "scale(1.08)",
-                        },
+                          transform: "scale(1.08)"
+                        }
                       }}
                     >
                       <Tooltip title="Copy">
@@ -532,7 +535,7 @@ function HeroSection() {
                           sx={{
                             p: "10px",
                             color: "#FE7D62",
-                            fontSize: "14px",
+                            fontSize: "14px"
                           }}
                           aria-label="search"
                         >
@@ -540,13 +543,13 @@ function HeroSection() {
                         </IconButton>
                       </Tooltip>
                     </InputAdornment>
-                  ),
+                  )
                 }}
                 inputProps={{
                   sx: {
                     color: "#FE7D62",
-                    fontSize: "18px",
-                  },
+                    fontSize: "18px"
+                  }
                 }}
               />
             )}
@@ -564,7 +567,7 @@ function HeroSection() {
                 width: "220px",
                 transform: "translate(-50%, -50%)",
                 border: "1px solid #FBF6F4",
-                aspectRatio: 1,
+                aspectRatio: 1
               }}
             >
               <img src={qrCode} style={{ width: "100%", height: "100%" }} />
@@ -579,7 +582,7 @@ function HeroSection() {
               bottom: "10px",
               right: "0px",
               width: "fit-content",
-              padding: "10px 20px",
+              padding: "10px 20px"
             }}
           >
             {alertMessage}
